+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Fixed
+
+* Correctly send multiple blob ids for multiple attachments.
+
 ## [0.1.0] — 2018-04-27
 
 ### Added